Create Array PAH 2.Missing & Repeating numbers
Saran is given an unsorted array of integers, where the array size is N and the elements in the array are integers in the range from 1 to N inclusive. 



However, there is one number missing from this range and one number appears twice in the array. 

Your task is to create a program to assist Saran that identifies and prints both the missing and the repeating numbers from the array.