Create function Q 3
Problem Statement



Raj is fascinated by different types of numbers and their properties. He wants to categorize a number, whether it is a perfect number or not. 



A perfect number is a positive integer that is equal to the sum of its proper divisors (excluding itself).



Help him write a program that takes a positive integer as input and outputs whether it is a perfect number or not.



Example



Input:

6

Output:

6 is a perfect number