Create class PAH 3
Rita wants to calculate the projected population of her town and needs to estimate the future population. 



She created a class called Population, which has the initial population, the annual growth increase, and the number of years as public attributes that are initialized to '0' using a default constructor. She used a method called calculate() to calculate the final population using the formula, Final Population = Initial population + (annual growth increase * years).



Assist Rita in completing the program.