Day 9 Q 4 Stack
Problem Statement



You are required to implement a C++ program that utilizes the Standard Template Library (STL) to perform operations on a stack. 



Write a program that takes the size of a stack as input, followed by the elements to be pushed onto the stack. The program then pops one element from the stack, pushes two additional elements, and finally prints the top element of the stack without popping it.