Create Day 9 Q 1
Emily needs a program to remove duplicates and find the sum of unique elements in an array using a container list.



The program prompts Emily to input the size of an array and then reads the array elements. It utilizes a list to store and sort the elements, keeping unique values. Finally, calculate and display the sum of these unique elements. If the array size is greater than 15, output '-1'.