Create Day 9 PAH Q 4-Lambda
Problem Statement



In a college project, students are tasked with creating a program to find the sum of two integers using lambda expressions. 



As part of their programming assignment, they must demonstrate their understanding of lambda functions by implementing a function that takes two integers as input and returns their sum. Your task is to guide the students.